WIP PoB Integration. Based on arevtur To experiment with it, set pob & build paths on evaluate.service.ts to something valid in your system, and the path in itemEval.js to where the lua binaries are.
When you use Ctrl+D to evaluate an item, it will compare it with your selected PoB Build.
The async databind is not working right, click on the evaluate dialog after a couple seconds
or change any value within it and it will update with the PoB tooltip comparison.

Everything is still WIP and hardcoded and help is appreciated. Things still left todo:
= Add PoB path and Build selection to options (along with other possible settings for this feature)
= All UI/UX work (does it belong in the evaluate dialog? is there a setting? light gray is awful)
= Convert itemEval.js and ScriptBase.js to typescript (if you change them, take care that they are git ignored)
= Fix the binding issue
= Add error handling, test with different item types (flasks are a bit broken for example)
= structure the code in proper services/components instead of everything out of place like it is

even with all the things left to fix it's already quite useful! have fun